Place `de` module at its final location
